+++ packages/core/lib/combo/confvars.ts
@@ -185,6 +185,7 @@ export const CONFVARS = [
   'ONLY_MM',
   'MM_KEEP_ITEMS_RESET',
   'MM_FAST_MASKS',
+  'OOT_PLANTED_BEANS'
 ] as const;
 
 export type Confvar = typeof CONFVARS[number];